Manh-12/4 - Fix bug price display wrong

+++ src/components/hooks/web3/useOwnedLeasingBooks.ts
@@ -33,7 +33,7 @@ export const hookFactory: OwnedLeasingBooksHookFactory =
               tokenId: item?.tokenId?.toNumber(),
               renter: item?.renter,
               amount: item?.amount?.toNumber(),
-              price: parseInt(ethers.utils.formatEther(item?.price)),
+              price: parseFloat(ethers.utils.formatEther(item?.price)),
               meta
             });
           } catch (err) {
